// src/common/enums.ts
export enum Color {
  black = "black",
  white = "white",
  gray = "gray",
  red = "red",
  green = "green",
  blue = "blue",
  yellow = "yellow",
  purple = "purple",
  pink = "pink",
}

export enum ParamType {
  string = "string",
  number = "number",
  gender = "gender",
  style = "style",
}

export enum StyleList {
  classic = "classic",
  official = "official",
}

export enum GenderList {
  male = "male",
  female = "female",
  unisex = "unisex",
}